Create README.md
Limited production exceeding project requirements for SWE3313.  The overall planning for this site was done as a group project.  All of the code structure was planned and coded by Russell Geary to become familiar with IntelliJ, Gluon and JavaFx.  A complete jar file will be available in the folders herein.